basic GET api working, need to do post
