--- src/SpikeTeam/RestBundle/Controller/SpikerController.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace SpikeTeam\RestBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-
-class SpikerController extends Controller
-{
-    /**
-     * Collection get action
-     * @var Request $request
-     * @return array
-     *
-     * @Rest\View()
-     */
-    public function cgetAction(Request $request)
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $entities = $em->getRepository('SpikeTeamRestBundle:Spiker')->findAll();
-
-        return array(
-            'entities' => $entities,
-        );
-    }
-}
