--- src/SpikeTeam/RestBundle/Entity/SpikerRepository.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace SpikeTeam\RestBundle\Entity;
-
-use Doctrine\ORM\EntityRepository;
-
-/**
- * SpikerRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class SpikerRepository extends EntityRepository
-{
-}
