--- src/SpikeTeam/RestBundle/Resources/config/routing.yml
@@ -1,3 +1,0 @@
-user:
-    type: rest
-    resource: SpikeTeam\RestBundle\Controller\SpikerController
