+++ src/SpikeTeam/UserBundle/Entity/User.php
@@ -22,27 +22,6 @@ class User extends BaseUser
      */
     protected $id;
 
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="name_first", type="string", length=60, nullable=true)
-     */
-    private $nameFirst = null;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="name_last", type="string", length=60, nullable=true)
-     */
-    private $nameLast = null;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column(name="phone_number", type="string", length=60, nullable=true)
-     */
-    private $phoneNumber = null;
-
     public function __construct()
     {
         parent::__construct();
@@ -58,72 +37,4 @@ class User extends BaseUser
         return $this->id;
     }
 
-    /**
-     * Set nameFirst
-     *
-     * @param string $nameFirst
-     * @return User
-     */
-    public function setNameFirst($nameFirst)
-    {
-        $this->nameFirst = $nameFirst;
-
-        return $this;
-    }
-
-    /**
-     * Get nameFirst
-     *
-     * @return string 
-     */
-    public function getNameFirst()
-    {
-        return $this->nameFirst;
-    }
-
-    /**
-     * Set nameLast
-     *
-     * @param string $nameLast
-     * @return User
-     */
-    public function setNameLast($nameLast)
-    {
-        $this->nameLast = $nameLast;
-
-        return $this;
-    }
-
-    /**
-     * Get nameLast
-     *
-     * @return string 
-     */
-    public function getNameLast()
-    {
-        return $this->nameLast;
-    }
-
-    /**
-     * Set phoneNumber
-     *
-     * @param string $phoneNumber
-     * @return User
-     */
-    public function setPhoneNumber($phoneNumber)
-    {
-        $this->phoneNumber = $phoneNumber;
-
-        return $this;
-    }
-
-    /**
-     * Get phoneNumber
-     *
-     * @return string 
-     */
-    public function getPhoneNumber()
-    {
-        return $this->phoneNumber;
-    }
 }
